Add commonmark extensions for goldmark_format
Since goldmark is based on commonmark (unlike blackfriday which isn't), we need to add all base commonmark extensions to markdown_strict.

Note that there are a couple of test failures in tests/testthat/test-md-document-hooks.txt. I don't believe thesre are actual failures but rather a result in changes in how pandoc generates markdown: (1) The underscore change is from the inclusion of `intraword_underscores`; (2) The table change is a result of a underlying change in how pandoc generates aligned table cells.